System Design/Lab2/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
bcrypt
python-dotenv
pyjwt
pymongo
motor
redis
aioredis